[LibOS] Explicitly set RDX
The comment says that the RDX register is set to NULL (0).
That said, the RDX is never set. Assuming that RDX is 0 is quite
optimistic. This commit explicitly sets RDX.
We use the movq instruction (instead of xor). The choice of
instruction is essential because we don't want to mess up rFLAGS.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>